3 - Update database schema

+++ solutions/2/DDL.sql
@@ -16,19 +16,19 @@ DROP TABLE IF EXISTS public.user_status;
 
 
 CREATE TABLE public.user_status (
-    user_status_id    SERIAL PRIMARY KEY,
+    user_status_id    BIGSERIAL PRIMARY KEY,
     status            VARCHAR(30) NOT NULL
 );
 
 
 CREATE TABLE public.relation_type (
-    relation_type_id  SERIAL PRIMARY KEY,
+    relation_type_id  BIGSERIAL PRIMARY KEY,
     relation_type     VARCHAR(30)
 );
 
 
 CREATE TABLE public.attachments (
-    attachment_id     SERIAL PRIMARY KEY,
+    attachment_id     BIGSERIAL PRIMARY KEY,
     location          VARCHAR(250) NOT NULL, -- location on hard disc
     mimetype          VARCHAR(50) NOT NULL,
     modification_date timestamp NOT NULL
@@ -36,12 +36,12 @@ CREATE TABLE public.attachments (
 
 
 CREATE TABLE public.users (
-    user_id           SERIAL PRIMARY KEY,
+    user_id           BIGSERIAL PRIMARY KEY,
     given_name        VARCHAR(100) NOT NULL,
     family_name       VARCHAR(100) NOT NULL,
     email             VARCHAR(100) NOT NULL,
-    status_id         INT,
-    attachment_id     INT,
+    status_id         BIGINT,
+    attachment_id     BIGINT,
 
     CONSTRAINT fk_attachment FOREIGN KEY (attachment_id) REFERENCES public.attachments(attachment_id) ON DELETE SET NULL,
     CONSTRAINT fk_status FOREIGN KEY (status_id) REFERENCES public.user_status(user_status_id) ON DELETE SET NULL
@@ -49,20 +49,20 @@ CREATE TABLE public.users (
 
 
 CREATE TABLE public.user_address (
-    user_address_id   SERIAL PRIMARY KEY,
-    user_id           INT NOT NULL,
+    user_address_id   BIGSERIAL PRIMARY KEY,
+    user_id           BIGINT NOT NULL,
 
     zip_code          VARCHAR(5) NOT NULL,
     street            VARCHAR(50) NOT NULL,
-    house_number      INT NOT NULL,
+    house_number      NUMERIC(5) NOT NULL,
     city              VARCHAR(30) NOT NULL,
 
     CONSTRAINT fk_user    FOREIGN KEY (user_id) REFERENCES public.users(user_id) ON DELETE CASCADE
 );
 
 CREATE TABLE public.user_contacts (
-    user_contact_id   SERIAL PRIMARY KEY,
-    user_id           INT NOT NULL,
+    user_contact_id   BIGSERIAL PRIMARY KEY,
+    user_id           BIGINT NOT NULL,
 
     mobile            VARCHAR(12),
     twitter           VARCHAR(40),
@@ -73,9 +73,9 @@ CREATE TABLE public.user_contacts (
 
 
 CREATE TABLE public.user_relations (
-    user1_id          INT NOT NULL,
-    user2_id          INT NOT NULL,
-    relation_id       INT,
+    user1_id          BIGINT NOT NULL,
+    user2_id          BIGINT NOT NULL,
+    relation_id       BIGINT,
 
     CONSTRAINT fk_user1 FOREIGN KEY (user1_id) REFERENCES public.users(user_id) ON DELETE CASCADE,
     CONSTRAINT fk_user2 FOREIGN KEY (user2_id) REFERENCES public.users(user_id) ON DELETE CASCADE,
@@ -83,13 +83,13 @@ CREATE TABLE public.user_relations (
 );
 
 CREATE TABLE public.roles (
-    role_id            SERIAL PRIMARY KEY,
+    role_id            BIGSERIAL PRIMARY KEY,
     role_name          VARCHAR(25) NOT NULL
 );
 
 CREATE TABLE public.role_user (
-    role_id            INT NOT NULL,
-    user_id            INT NOT NULL,
+    role_id            BIGINT NOT NULL,
+    user_id            BIGINT NOT NULL,
 
     CONSTRAINT fk_user_id FOREIGN KEY (user_id) REFERENCES public.users (user_id) ON DELETE CASCADE,
     CONSTRAINT fk_role_id FOREIGN KEY (role_id) REFERENCES public.roles (role_id) ON DELETE CASCADE
@@ -100,7 +100,7 @@ CREATE TABLE public.role_user (
 
 CREATE TABLE public.meetings (
     meeting_id        SERIAL PRIMARY KEY,
-    organizer_id      INT NOT NULL,
+    organizer_id      BIGINT NOT NULL,
     note              VARCHAR(100),
     description       TEXT,
     location          VARCHAR(75) NOT NULL,
@@ -111,8 +111,8 @@ CREATE TABLE public.meetings (
 );
 
 CREATE TABLE public.meeting_invitations (
-    meeting_id        INT NOT NULL,
-    user_id           INT NOT NULL,
+    meeting_id        BIGINT NOT NULL,
+    user_id           BIGINT NOT NULL,
 
     CONSTRAINT fk_meeting FOREIGN KEY (meeting_id) REFERENCES public.meetings(meeting_id) ON DELETE CASCADE,
     CONSTRAINT fk_user FOREIGN KEY (user_id) REFERENCES public.users(user_id) ON DELETE CASCADE
